fix(combat): PE-canon re-label — aberrant combat resource PE->SG, close XP collision (#2527 verdict A)
Per #2527 (master-dd verdict A): PE = campaign XP (26-ECONOMY_CANONICAL), not a combat resource. Closes the collision where PHASEC combat-PE (#2522/#2526) wrote unit.pe (the same field forms/evolution uses for XP). Combat no longer writes unit.pe: removed first_kill_pe_bonus combat earn + #2526 base PE earn (applyBaseAbilityResourceEarn). aberrant_overdrive cost_pe->cost_sg (data, decorative). unit.pe stays as read-only campaign-XP pass-through. first_kill_pe_bonus/minion_kill_pe_bonus = campaign-XP perks (combat-unwired). Finding: cost_sg scales are inconsistent (3..100 vs SG pool max 3) -> a real combat-cost gate is a separate economy-alignment follow-up (dropped the gate to avoid breaking surge_aoe). TDD; 96/96 progression+abilityExecutor, AI 495/495, datasets valid.

Coding-Agent: claude-opus-4.8
Trace-Id: 019e8364-6ff7-7c04-b3f5-e25c0877c8c0
